Enable user specifiable timer for performance runs
[reviewed by kyle]

Previously the presence of PERFTIMEEXEC or .perftimeexec file indicated that a
performance test should be run with time -p and then the timings were picked
out with the perfkeys functionality. However time -p is not a very precise
timer so it's not very useful for timing things like non-user code startup
(which is under .001 seconds for fifo.)

This patch changes it so that the PERFTIMEXEC or .perftimeexec files contains
the name of a timer to use. For instance you could have a PERFTIMEXEC with:

defaultTimer

And it would use util/test/timers/defaultTimer to time the program execution.

The timer must be located in $CHPL_HOME/util/test/timers. If the specified
timer cannot be found or the perftimeexec file was badly formatted it attempts
to use the defaultTimer, and if that too fails, it reverts to time -p and
prints out an error. The timer is also responsible for actually running the
executable as well.

The current timers are defaultTimer (a simple wrapper for time -p) and
highPrecisionTimer (very precise cross-platform timer provided by kyleb, which
uses python's timeit module.) It only captures real time, and not sys or user
though.

The rest of the changes just update the notes in start_test and replace all the
existing PERFTIMEEXEC and perftimeexec content with defaultTimer.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23803 3a8e244f-b0f2-452b-bcba-4c88e055c3ca