Add a no-op performance test
[reviewed by kyle]

The no-op test is intended to measure the amount of time spent setting up
non-user code. It's been observed that this is non-insignificant for qthreads,
and it will be interesting to see the results for multi-locale runs as well. I
added a new suite for startup tracking with the thinking that we may want to
start tracking startup communications once we have multi-locale perf testing
running.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23804 3a8e244f-b0f2-452b-bcba-4c88e055c3ca