Add tests for the -f execution-time options.
[Not reviewed.]

Add tests for the execution-time -f option, which specifies a file from
which config values are to be taken.  There is a test for "-f<file>"
(both parts in one command line option) and one for "-f <file>" (the two
parts in separate options).

These pass now, but would have failed prior to r23817.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23820 3a8e244f-b0f2-452b-bcba-4c88e055c3ca