Cleanup of type queries for primitive casts in ChapelBase.chpl
[reviewed by Vass]

Context: As Vass has been strategizing about cleaning up type queries,
as a background task, I started digging into cases where code was
using _isPrimitiveType() to understand what was really being asked, in
order to either improve the definition of _isPrimitiveType() to better
match the spec, or refine its name, or retire it.

This commit takes a step along that path by replacing uses of it to
define casts between types that are implemented by the compiler via
primitives (bools, ints, uints, reals).  It turns out that we don't
teach the compiler how to cast for imaginary and complex types (which
probably keeps the compiler simpler) and implement these as module
code instead.  This is the reason that the traditional
_isPrimitiveType() query excluded those types even though they are
"primitive" by the language's definition.

To address this issue, I cloned _isPrimitiveType() into ChapelBase,
named it chpl_typeSupportsPrimitiveCast() to make its name more
specific to its use, and protect it via chpl_ rather than just a
leading underscore.  This should improve clarity and move us one
step closer to retiring or redefining _isPrimitiveType().




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23821 3a8e244f-b0f2-452b-bcba-4c88e055c3ca