Make the filerator test more portable
For some reason that I haven't been able to determine, the return type of
readdir() is sometimes dirent64* rather than dirent* as all the man pages and
documentation that I can find suggest.  This happens with prgenv-pgi and
prgenv-cray compilers for example.  Here, I'm guarding our typedef of
'direntptr' for Chapel's use with a similar #ifndef structure as the
/usr/lib/dirent.h file to make this test work on our current systems.  This is
likely be something that we need to understand and resolve better before
putting a filerator into our standard libraries, though... :(



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23824 3a8e244f-b0f2-452b-bcba-4c88e055c3ca