Committing a symlink to aprun-utils.c and removing the copy in the Makefile
[reviewed by Thomas]

This commit checks in a symlink of the file aprun-utils.c in the
pbs-aprun directory.  It also removes the copying of that file as well
as fixes a type-o.

The copying of aprun-utils.c into the pbs-aprun directory was
problematic for the dependence checking, but I didn't notice b/c I
already had a rule for building aprun-utils.c in my repository at the
time.

Apparently SVN symlinks do not work on Window systems, but there are
no Windows systems than support aprun, so I'm not going to worry about
it.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23825 3a8e244f-b0f2-452b-bcba-4c88e055c3ca