Update setting hwpar for the nemesis scheduler
[ok'd by greg]

Qthreads doesn't do smart things with QT_HWPAR for different schedulers (at
least on systems for numa domains.)

For instance on chap04 (two numa domains) if you set QT_HWPAR=8, qthreads will
use 2 shepherds with 4 workers each with the nemesis scheduler, even though you
need 8 shepherds and one worker per (since the nemesis scheduler only permits
single threaded shepherds.)

This is a hacky way to get the right behavior that is intended only to test out
the performance of the nemesis scheduler on chap04. It shouldn't be in here
for more than a week.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23836 3a8e244f-b0f2-452b-bcba-4c88e055c3ca