Make paratest look in certain directories where it should and it didn't.
[reviewed by Sung]

Turns out that paratest.server skips some directories that have no
.chpl files in them and that start_test doesn't skip.

As a result, when someone runs parallel testing, certain directories
with C tests only are skipped. If there are errors in those
directories, they are not reported to the person running the
tests. Those errors are then picked by the nightly scripts.

Directories are filtered by chpl_files() in paratest.server and by a
combination of conditions in start_test. This patch adjusts the former
to (supposedly) match the latter.

In other words, The logic around *.test.c is the same in paratest and
start_test, AFAICT. The mismatch that this change addresses happens
when the directory has no .chpl or .test.c files, but has a sub_test.

While there:

* In paratest.server react to NOTEST, rather than *NOTEST, to match start_test.

* Add a cross-reference comment to the corresponding condition in start_test. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23851 3a8e244f-b0f2-452b-bcba-4c88e055c3ca