Don't force QT_WORKER_UNIT=core with the nemesis thread scheduler.
[Reviewed by Elliot.]

We limit the #workers to no more than the #PUs (hyperthreads) in
qthreads.  But we also tell qthreads to distribute the workers onto
processor cores, not PUs.  This creates a conflict when the #workers is
greater than the #cores.  The regular qthreads thread scheduler is okay
with this, but the nemesis scheduler gripes about it.  So, if we're
using the nemesis scheduler, don't force worker distribution onto the
cores, just let qthreads do what it wants in this case.

This is an adjustment to r23836, and should be removed when that commit
is removed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23864 3a8e244f-b0f2-452b-bcba-4c88e055c3ca