Fixing a PGI regression
[stenographing for Kyle on behalf of Michael]

PGI was failing on this C test because one of the operations in use
isn't supported by intrinsics, and the lookup table to support it is
in a .c rather than .h file.  Added the .c file to the .compopts which
provides the table and should make this test pass on the next PGI
compile.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23868 3a8e244f-b0f2-452b-bcba-4c88e055c3ca