Seeing what happens with .perfexecenv removed
[discussed with Lydia]

In discussions today, we've been having trouble precisely why this
.perfexecenv was used, and our best guess was that Lydia thought she
saw performance improvements with it.  To that end, we wanted to 
remove it for a few nights, to see if that's the case; the ultimate
goal being to reduce the special settings that tests require (in
general and for our shootout entry).

Greg points out that a benefit of this file is that it'll turn a
deadlock into an error message if the system resources are exceeded;
but given that the testing system has timeouts and the setting isn't
helpful outside of the performance testing system (and we haven't seen
systems on which we hit those limits), I'd prefer not to have the
file, if there are no other downsides.

If, after some performance testing, we see a change, we can discuss
whether that's surprising, and should reinstate it.  I'm open to
more discussion on Greg's take as well, but not enough to not see
what the performance impacts are for a few days.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23869 3a8e244f-b0f2-452b-bcba-4c88e055c3ca