The overarching goal of this commit is to set the framework up for multiple (simultaneous) file systems, as well as fix some small errors that I found recently while compiling on my Mac machine.
This commit does a couple things:
(1) It changes the make system for auxFilesys to support multiple types of file
    systems. We now set variables in the makefiles, as opposed to choosing which
    makefiles to include.
(2) It fixes an error in which "EREOMTEIO" is not found in errno.h on Mac machines.
    This is fixed by simple ifdefs in the file.
(3) It fixes an error when compiling with clang. We would previously get "treating
    'c' input as 'c++' when in C++ mode, this behavior is deprecated". This is fixed
    by renaming the extensions of the files so that Clang will treat them as C++
    files (.c -> .cc)
(4) We remove the "none" directory in the "auxfilesys" directory. The stub file is
    now moved into "hdfs".
(5) We remove the superfluous library linkages (clang was complaining about this).