Changing license from 3-clause BSD to Apache
As discussed over much time within Cray and on the Chapel mailing
lists, we are moving to the Apache License (and, in a separate
commit, contributor agreements) in order to move away from our
previous homegrown contributor agreements while getting a similar
level of protection from a standard form.  Hopefully this will
make the lawyers of the world happy without making any users sad.