Move stack declarations for aprun argument list to global scope.
These declarations were previously at the global scope, but I had
moved them on to the stack in my previous commit (r23802 in the SVN
world) because my original design was to copy them before returning.
I decided to keep the design that was originally there before my
commit, but forgot to move the local declarations back to the global
scope.

This worked in local testing likely because I did not build optimized.