Futurizing these tests.
As the functions in question currently stand, a value will always be returned,
but the compiler is under the impression this is not the case and so is
generating an error.  With proper analysis, these two cases should pass, so the
issue is now a future for both of them.