This fixes the problem with c_string not being initialized.
c_string is not really an extern type, but I had it marked as such after creation to
prevent code generation.  FLAG_NO_CODEGEN is a better choice for this.

I updated the usage on a few other types as well, as appropriate.  _file_list is genuinely
defined externally, so it still needs FLAG_EXTERN.