Remove double definition of numLocales config var.
I inlined ChapelNumLocales in the localeModel files to work around a circular reference
problem in the modules.  I removed "use ChapelNumLocales" from ChapelStandard, but
neglected to remove it from the LocaleModel files as well.  The double-defintion made the
variable show up twice in help text, which made a number of help-releated tests fail.

With this fix, the failure count is down to 54.