Adds the _defaultOf generation for tuples.
Not all type information or substitutions for tuples are known by the time
buildDefaultFunctions comes around, but it needs to have a _defaultOf method
around to instantiate by the time functionResolution rolls in for generic
instantiation.  So create a stub in the former and fill in the body in the
latter.