Another attempt to get freeNil.chpl working.
The FLAG_EXTERN applied to dtOpaque was a lazy hack to get the compiler to generate the
assignment operator for it.  But FLAG_EXTERN now means that no initialization is
supplied.  (How can we know how to initialize an extern type?)

So a better solution is to treat dtOpaque as a fundamental type and supply a "real"
initializer for it.

This pushes the failure count down to 35.

[tested on / (std)]