Added a bug future whereby the compiler crashes when a task intent's ref clause includes a reference to a field.
From the .future: instead an error should be generated.
That's because fields are not subject to forall intents.
Even fields of a record.

Also, my anticipated implementation of forall intents will likely hit
the same issue.

P.S. This should really be 8 distinct tests, to ensure that an error
is generated for each case. The param 'if' I am using does not guarantee
that - because the compiler may generate the error prior to resolution.
So ideally split this into 6 tests once the bug is fixed,
the add two for forall intents when those are working.