Reinstate creation of the _statementLevelSymbol call.
I had bridged out the section of code that creates the call to _statementLevelSymbol() for
SymExprs that appear at the statement level.  The verification failure that was earlier
caused by that insertion appears to have been cured by intervening changes on this branch.

Also reverted all the test sources, where I worked around the missing call to
_statementLevelSymbol() by calling the read methods directly for sync and single
variables.

Now passes all tests in the standard configuration.