Try out the sherwood scheduler again for qthreads with condwait-queue
I want to see what the performance for the sherwood scheduler is like again now
that we have condwait-queue enabled. It would be great if we could use the
sherwood scheduler instead of the nemesis. I don't think that's going to be the
case and I'm pretty sure there's more overhead to using the sherwood scheduler
over nemesis, but I want to confirm this with some testing tonight.

Like before this disables work stealing since that had a negative performance
impact on a lot of tests, but this makes disabling work stealing dependent on
an environment variable instead of forcing it like I did last time.