Get users/ferguson/recordConstructor test case working.
Removed recordConstructor.future to force correct operation of this test case.

In resolvePrimInit(), replace the hard link to the defaultInitializer function with a call
that is generated by generateConcreteConstructorCall().  This new function uses the
substitution map to call the constructor for the given type by name, using the
substitutions to provide the and type and param arguments that were supplied to resolve an
originally generic type to the given concrete type.

The compilerError call in file.file was commented out.  Apparently some old code still
resolves to this, and (now that default constructors work), the code will be called if a
"file" object is created *anywhere*.  TODO: Remove mention of the old-style "file" object
from the modules, if possible.