Fix some tests in types/records/sungeun
Updated classInRecord.good to match the behavior on the branch.  Examining the source
code, it appears that it is correct for the default-constructor of myR to be called even
when no initializer is provided in the declaration.  That means "c = new myC();" will
always be called when a record is initialized.  Hence, no cases remain where the "c" field
in a myR record can initially be nil.  The revised .good file shows that the "c" field is
now always (correctly) initialized.

Futurized the test case constructor2.  Normalization currently converts declaration
constructs that are intended to represent copy-initialization with default-initialization
followed by assignment.  Details appear in the added .future file.  Fixing the underlying
problem is well beyond the scope of the current effort.