Handle the special case where a generated constructor call involves runtime type arguments
As currently implemented, runtime type handling will not create a SymExpr containing a
TypeSymbol referring to a type flagged with FLAG_HAS_RUNTIME_TYPE into a _RuntimeTypeInfo
instance.  To coerce insertRuntimeInitTemps() into doing its job, I added the function
fixupRuntimeTypeArguments() and call this with the new default constructor call after it
has been inserted into the tree.

For each actual argument that names a runtime type, fixupRuntimeTypeArguments() inserts a
type temp and initializes it through a call to PRIM_INIT (passing that type).  This
provides runtime type handling with AST that conforms to its expectations.