Change qthreads topology/affinity setup to work with nemesis
qthreads seems to setup the number of workers and workers per shepherd assuming
the shwerwood scheduler is being used when configured with hwloc. This changes
some of the logic to work with the nemesis scheduler. I have the patch out to
Dylan as well and will make changes that he suggests. I may go with this for a
little while if I don't hear back from him just to get some testing in.