Get two test cases infer_field10 and infer_field11 working again.
Creation of the default constructor call initially assumed that all substitutions are
bound to a param or type argument.  That is true in the signature of the default type
constructor, but not necessarily so in the signature of the default constructor.  In the
latter, the type of an argument may be marked as dtAny, so a value is expected and the
formal argument type is inferred from the type of the actual argument.

Formal arguments flagged with FLAG_TYPE_VARIABLE are removed in pruneResolvedTree() so it
is clear that a constructor argument expecting a value will not have that flag.

However, to avoid having to match up actuals with formals*, I used the FLAG_GENERIC flag
to indicate that an argument in the default type constructor that is flagged with
FLAG_TYPE_VARIABLE will appear in the default constructor without that flag.
* The default constructor call is generated before it is resolved, so that constructor
calls can be resolved by name.

###compiler/include/flags_list.h
Added a comment to explain the expanded use of FLAG_GENERIC.

###compiler/passes/scopeResolve.cpp
Arguments to the default type constructor are now tagged with FLAG_GENERIC if they are
marked as FLAG_TYPE_VARIABLE but the corresponding field is not.

###compiler/resolution/functionResolution.cpp
Added a new clause that inserts a PRIM_INIT expression to convert a type symbol into a
value when the constructor is being called in a way that performs type inference from a
type.  This is the case described above, where a formal argument in the default type
constructor is flagged with FLAG_TYPE_VARIABLE but the corresponding formal in the default
constructor is not -- which corresponds exactly to the case where that formal in the
default type constructor is flagged with FLAG_GENERIC.

Added a new function flattenAndResolveArgs() to create call temps for the PRIM_INIT calls
inserted into the argument list of the created constructor call, and to resolve them.
This is called before fixupRuntimeTypeArguments().