Resolve 4 more failures due to previous fix.
The test classes/hilde/explainInstantiation is sensitive to resolution order, and the
inserted constructor calls can change that order.  The .good file was updated to reflect
the current output.

The inserted conversion of a type symbol into a value (through PRIM_INIT) is not necessary
if the substituted value is already a param.  Piping it through PRIM_INIT erases its value
and (somewhat surprisingly) its param-ness.  The latter change causes an actual/formal
mismatch.

To cure this problem, I inserted a test so the type-to-value conversion is only inserted
if the substituted value is actually a type.