Updating test scripts to make release-over-release work again
[this is Brad, gathering release-over-release perforamnce as chapelu]

Recent changes to start_test and sub_test broke release-over-release
performance because they based paths off of $CHPL_HOME rather than
$CHPL_TEST_UTIL_DIR (which release-over-release requires because
CHPL_HOME is set to the release being tested, but the test directory
and util directories are used from the current release.

Some other changes were also required because we used to run the
util/chplenv scripts from the release itself in the testing system,
but now that the scripts have changed names, we can't do that for the
releases.  To that end, I made $CHPL_TEST_UTIL_DIR point to util/
where previously it pointed to util/test.  That way we can either
get to util/test or to util/chplenv.

Finally, this commit includes a few changes that have been sitting
around in testReleasesPerformance for awhile -- the main one being
that we no longer run against the current repository since we're now
doing that on a nightly basis and splicing those results in.